feat: Add throttling to heartbeat endpoint
This commit introduces a throttling mechanism to the heartbeat endpoint.
A new `ThrottleState` struct is added to the `AppState` to keep track of
the last request time for each user-channel pair. If a user tries to send
a heartbeat for the same channel within a minute, a 429 Too Many Requests
response is returned. The throttling state is shared across all requests
using an `Arc<Mutex<ThrottleState>>`.

Signed-off-by: Daniel Boll <danielboll.academico@gmail.com>