feat(throttle): cap the requests for the heartbeat endpoint (#4)
## First Approach

This commit introduces a throttling mechanism to the heartbeat endpoint.
A new `ThrottleState` struct is added to the `AppState` to keep track of
the last request time for each user-channel pair. If a user tries to
send
a heartbeat for the same channel within a minute, a 429 Too Many
Requests
response is returned. The throttling state is shared across all requests
using an `Arc<Mutex<ThrottleState>>`.

Signed-off-by: Daniel Boll <danielboll.academico@gmail.com>

## Second Approach

After realizing that would be quite similar to use ScyllaDB for this
task, we did a new table for Throttling using TTL in the insertion of
the query.

The partition key is made of:
- `request_uri`: /some/route
- `user_id`: user related to the Bearer Token sent in the request
- `content`: a random string that will be used to differentiate this
request from the others.

```sql
CREATE TABLE twitch.throttle_v1 (
    uri text,
    user_id int,
    content text,
    updated_at timestamp,
    PRIMARY KEY ((uri, user_id, content), updated_at)
) WITH CLUSTERING ORDER BY (updated_at ASC)


INSERT INTO throttle_v1 (uri, user_id, content, updated_at) VALUES (?, ?, ?, ?) USING TTL ?
```

Each request can have a dynamic TTL based on the user necessity: 

```rust
impl Throttle {
  pub async fn insert_throttle(&self, connection: &CachingSession, ttl: i32) -> anyhow::Result<()> {
    let query = Query::new(INSERT_THROTTLE_WITH_TTL);

    connection
      .execute(query, (
          &self.uri, 
          &self.user_id,
          &self.content,
          &self.updated_at,
          ttl,
        ),
      )
      .await?;

    Ok(())
  }
}
```

---------

Signed-off-by: Daniel Boll <danielboll.academico@gmail.com>
Co-authored-by: danielhe4rt <hey@danielheart.dev>